Separated app from server listen logic for supertest, removed testing logic from db setup
