+++ server/server.js
@@ -1,29 +1,6 @@
-import dotenv from 'dotenv';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import express from 'express';
-import cors from 'cors';
-import router from './routes/router.js';
-
-const app = express();
-
-// Resolve the project root .env path
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-dotenv.config({ path: path.join(__dirname, '../.env') });
+import { app } from './app.js';
 
 const PORT = process.env.SERVER_PORT || 3000;
-
-app.use(express.json());
-app.use(cors({ origin: true, credentials: true }));
-app.use(express.urlencoded({ extended: true }));
-
-app.use('/', router);
-
-app.use((err, req, res, next) => {
-  console.error(err.stack);
-  res.status(500).send('Something broke!');
-});
 
 app.listen(PORT, () => {
   console.log(`Server is running on http://localhost:${PORT}`);
